separate traits and types
