+++ pallets/auctions/src/lib.rs
@@ -3,7 +3,6 @@
 #![allow(clippy::upper_case_acronyms)]
 
 // Used for encoding/decoding into scale
-use codec::{Decode, Encode};
 use frame_support::{
 	dispatch::{DispatchError, DispatchResult},
 	ensure,
@@ -11,7 +10,7 @@ use frame_support::{
 		tokens::nonfungibles::Inspect, Currency, ExistenceRequirement, LockIdentifier, LockableCurrency,
 		WithdrawReasons,
 	},
-	BoundedVec, Parameter,
+	Parameter,
 };
 use frame_system::{ensure_signed, RawOrigin};
 use sp_runtime::{
@@ -22,9 +21,11 @@ use sp_runtime::{
 	Permill,
 };
 use sp_std::result;
+pub use traits::*;
 use weights::WeightInfo;
 
 mod benchmarking;
+pub mod traits;
 pub mod weights;
 
 #[cfg(test)]
@@ -41,115 +42,6 @@ const BID_STEP_PERC: u32 = 10;
 const BID_ADD_BLOCKS: u32 = 10;
 /// Minimal auction duration
 const MIN_AUCTION_DUR: u32 = 10;
-
-#[derive(Encode, Decode, Clone, PartialEq, Eq)]
-pub enum Auction<T: Config> {
-	English(EnglishAuction<T>),
-}
-
-#[derive(Encode, Decode, Clone, PartialEq, Eq)]
-pub struct EnglishAuction<T: Config> {
-	pub general_data: GeneralAuctionDataOf<T>,
-	pub specific_data: EnglishAuctionData<T>,
-}
-
-#[derive(Encode, Decode, Clone, PartialEq, Eq)]
-pub struct EnglishAuctionData<T: Config> {
-	pub reserve_price: BalanceOf<T>,
-}
-
-#[derive(Encode, Decode, Clone, PartialEq, Eq)]
-pub struct GeneralAuctionData<AccountId, Balance, BlockNumber, NftClassId, NftTokenId, BoundedVec> {
-	pub name: BoundedVec,
-	pub last_bid: Option<(AccountId, Balance)>,
-	pub next_bid_min: Balance,
-	pub start: BlockNumber,
-	pub end: BlockNumber,
-	pub owner: AccountId,
-	pub token: (NftClassId, NftTokenId),
-}
-
-/// Define type aliases for better readability
-pub type GeneralAuctionDataOf<T> = GeneralAuctionData<
-	<T as frame_system::Config>::AccountId,
-	BalanceOf<T>,
-	<T as frame_system::Config>::BlockNumber,
-	<T as pallet_uniques::Config>::ClassId,
-	<T as pallet_uniques::Config>::InstanceId,
-	BoundedVec<u8, <T as pallet_uniques::Config>::StringLimit>,
->;
-
-pub type BalanceOf<T> = <<T as Config>::Currency as Currency<<T as frame_system::Config>::AccountId>>::Balance;
-
-impl<T: Config> sp_std::fmt::Debug for Auction<T> {
-	fn fmt(&self, f: &mut sp_std::fmt::Formatter) -> sp_std::fmt::Result {
-		write!(f, "Auction")
-	}
-}
-pub trait NftAuction<AccountId, AuctionId, BalanceOf, NftAuction> {
-	fn bid(bidder: AccountId, auction_id: AuctionId, value: BalanceOf) -> DispatchResult;
-
-	fn close(&self) -> DispatchResult;
-}
-
-impl<T: Config> NftAuction<T::AccountId, T::AuctionId, BalanceOf<T>, Auction<T>> for EnglishAuction<T> {
-	fn bid(bidder: T::AccountId, auction_id: T::AuctionId, value: BalanceOf<T>) -> DispatchResult {
-		<Auctions<T>>::try_mutate(auction_id, |maybe_auction| -> DispatchResult {
-			let auction = maybe_auction.as_mut().ok_or(Error::<T>::AuctionNotExist)?;
-			let Auction::English(data) = auction;
-				// Lock / Unlock funds
-				if let Some(ref current_bid) = data.general_data.last_bid {
-					<T as crate::Config>::Currency::remove_lock(AUCTION_LOCK_ID, &current_bid.0);
-				}
-				<T as crate::Config>::Currency::set_lock(AUCTION_LOCK_ID, &bidder, value, WithdrawReasons::all());
-
-				data.general_data.last_bid = Some((bidder, value));
-				// Set next minimal bid
-				let bid_step = Permill::from_percent(BID_STEP_PERC).mul_floor(value);
-				data.general_data.next_bid_min = value.checked_add(&bid_step).ok_or(Error::<T>::BidOverflow)?;
-
-				// Avoid auction sniping
-				let block_number = <frame_system::Pallet<T>>::block_number();
-				let time_left = data
-					.general_data
-					.end
-					.checked_sub(&block_number)
-					.ok_or(Error::<T>::TimeUnderflow)?;
-				if time_left < BID_ADD_BLOCKS.into() {
-					data.general_data.end = block_number + BID_ADD_BLOCKS.into();
-				}
-			
-
-			Ok(())
-		})
-	}
-
-	fn close(&self) -> DispatchResult {
-		pallet_uniques::Pallet::<T>::thaw(
-			RawOrigin::Signed(self.general_data.owner.clone()).into(),
-			self.general_data.token.0,
-			self.general_data.token.1,
-		)
-		.unwrap_or_default();
-		// there is a bid so let's determine a winner and transfer tokens
-		if let Some(ref winner) = self.general_data.last_bid {
-			let dest = T::Lookup::unlookup(winner.0.clone());
-			let source = T::Origin::from(frame_system::RawOrigin::Signed(self.general_data.owner.clone()));
-			pallet_nft::Pallet::<T>::transfer(source, self.general_data.token.0, self.general_data.token.1, dest)
-				.unwrap_or_default();
-			<T as crate::Config>::Currency::remove_lock(AUCTION_LOCK_ID, &winner.0);
-			<<T as crate::Config>::Currency as Currency<T::AccountId>>::transfer(
-				&winner.0,
-				&self.general_data.owner,
-				winner.1,
-				ExistenceRequirement::KeepAlive,
-			)
-			.unwrap_or_default();
-		}
-
-		Ok(())
-	}
-}
 
 pub use pallet::*;
 
@@ -336,7 +228,7 @@ pub mod pallet {
 }
 
 impl<T: Config> Pallet<T> {
-	fn validate_general_data(general_data: &GeneralAuctionDataOf<T>) -> DispatchResult {
+	fn validate_general_data(general_data: &GeneralAuctionData<T>) -> DispatchResult {
 		let current_block_number = frame_system::Pallet::<T>::block_number();
 		ensure!(
 			general_data.start >= current_block_number,
@@ -357,7 +249,7 @@ impl<T: Config> Pallet<T> {
 		Ok(())
 	}
 
-	fn validate_create(general_data: &GeneralAuctionDataOf<T>) -> DispatchResult {
+	fn validate_create(general_data: &GeneralAuctionData<T>) -> DispatchResult {
 		let is_transferrable = pallet_uniques::Pallet::<T>::can_transfer(&general_data.token.0, &general_data.token.1);
 		ensure!(is_transferrable, Error::<T>::TokenFrozen);
 		Ok(())
@@ -366,7 +258,7 @@ impl<T: Config> Pallet<T> {
 	fn handle_create(
 		sender: <T>::AccountId,
 		auction: &Auction<T>,
-		general_data: &GeneralAuctionDataOf<T>,
+		general_data: &GeneralAuctionData<T>,
 	) -> DispatchResult {
 		let auction_id = <NextAuctionId<T>>::try_mutate(|next_id| -> result::Result<<T>::AuctionId, DispatchError> {
 			let current_id = *next_id;
@@ -391,7 +283,7 @@ impl<T: Config> Pallet<T> {
 		Ok(())
 	}
 
-	fn validate_update(sender: <T>::AccountId, general_data: &GeneralAuctionDataOf<T>) -> DispatchResult {
+	fn validate_update(sender: <T>::AccountId, general_data: &GeneralAuctionData<T>) -> DispatchResult {
 		ensure!(general_data.owner == sender, Error::<T>::NotAuctionOwner);
 
 		let current_block_number = frame_system::Pallet::<T>::block_number();
@@ -407,7 +299,7 @@ impl<T: Config> Pallet<T> {
 		sender: <T>::AccountId,
 		auction_id: T::AuctionId,
 		updated_auction: Auction<T>,
-		general_data: &GeneralAuctionDataOf<T>,
+		general_data: &GeneralAuctionData<T>,
 	) -> DispatchResult {
 		<Auctions<T>>::try_mutate(auction_id, |auction_result| -> DispatchResult {
 			if let Some(_auction) = auction_result {
@@ -420,7 +312,7 @@ impl<T: Config> Pallet<T> {
 		})
 	}
 
-	fn handle_destroy(auction_id: T::AuctionId, general_data: &GeneralAuctionDataOf<T>) -> DispatchResult {
+	fn handle_destroy(auction_id: T::AuctionId, general_data: &GeneralAuctionData<T>) -> DispatchResult {
 		pallet_uniques::Pallet::<T>::thaw(
 			RawOrigin::Signed(general_data.owner.clone()).into(),
 			general_data.token.0,
@@ -437,7 +329,7 @@ impl<T: Config> Pallet<T> {
 
 	fn validate_bid(
 		bidder: &<T>::AccountId,
-		general_auction_data: &GeneralAuctionDataOf<T>,
+		general_auction_data: &GeneralAuctionData<T>,
 		value: BalanceOf<T>,
 	) -> DispatchResult {
 		let block_number = <frame_system::Pallet<T>>::block_number();
@@ -467,6 +359,65 @@ impl<T: Config> Pallet<T> {
 					EnglishAuction::<T>::close(data)?;
 				}
 			}
+		}
+
+		Ok(())
+	}
+}
+
+impl<T: Config> NftAuction<T::AccountId, T::AuctionId, BalanceOf<T>, Auction<T>> for EnglishAuction<T> {
+	fn bid(bidder: T::AccountId, auction_id: T::AuctionId, value: BalanceOf<T>) -> DispatchResult {
+		<Auctions<T>>::try_mutate(auction_id, |maybe_auction| -> DispatchResult {
+			let auction = maybe_auction.as_mut().ok_or(Error::<T>::AuctionNotExist)?;
+			let Auction::English(data) = auction;
+				// Lock / Unlock funds
+				if let Some(ref current_bid) = data.general_data.last_bid {
+					<T as crate::Config>::Currency::remove_lock(AUCTION_LOCK_ID, &current_bid.0);
+				}
+				<T as crate::Config>::Currency::set_lock(AUCTION_LOCK_ID, &bidder, value, WithdrawReasons::all());
+
+				data.general_data.last_bid = Some((bidder, value));
+				// Set next minimal bid
+				let bid_step = Permill::from_percent(BID_STEP_PERC).mul_floor(value);
+				data.general_data.next_bid_min = value.checked_add(&bid_step).ok_or(Error::<T>::BidOverflow)?;
+
+				// Avoid auction sniping
+				let block_number = <frame_system::Pallet<T>>::block_number();
+				let time_left = data
+					.general_data
+					.end
+					.checked_sub(&block_number)
+					.ok_or(Error::<T>::TimeUnderflow)?;
+				if time_left < BID_ADD_BLOCKS.into() {
+					data.general_data.end = block_number + BID_ADD_BLOCKS.into();
+				}
+			
+
+			Ok(())
+		})
+	}
+
+	fn close(&self) -> DispatchResult {
+		pallet_uniques::Pallet::<T>::thaw(
+			RawOrigin::Signed(self.general_data.owner.clone()).into(),
+			self.general_data.token.0,
+			self.general_data.token.1,
+		)
+		.unwrap_or_default();
+		// there is a bid so let's determine a winner and transfer tokens
+		if let Some(ref winner) = self.general_data.last_bid {
+			let dest = T::Lookup::unlookup(winner.0.clone());
+			let source = T::Origin::from(frame_system::RawOrigin::Signed(self.general_data.owner.clone()));
+			pallet_nft::Pallet::<T>::transfer(source, self.general_data.token.0, self.general_data.token.1, dest)
+				.unwrap_or_default();
+			<T as crate::Config>::Currency::remove_lock(AUCTION_LOCK_ID, &winner.0);
+			<<T as crate::Config>::Currency as Currency<T::AccountId>>::transfer(
+				&winner.0,
+				&self.general_data.owner,
+				winner.1,
+				ExistenceRequirement::KeepAlive,
+			)
+			.unwrap_or_default();
 		}
 
 		Ok(())
