+++ pallets/auctions/src/tests.rs
@@ -1,6 +1,6 @@
 use super::*;
 use crate::mock::*;
-use frame_support::{assert_noop, assert_ok};
+use frame_support::{assert_noop, assert_ok, BoundedVec};
 use sp_std::convert::TryInto;
 
 pub type AuctionsModule = Pallet<Test>;
